feat: alinhamento nativo, modo roteiro, tag drill-down, novos themes
- TopToolbar: botão de alinhamento (L/C/R/J) com popup de escopo
  (esse doc / todos / manuscrito / gavetas); ícones align-*.svg
- TopToolbar: botões de tamanho e espaçamento mais compactos
- Alinhamento reflete doc atual via cursorPositionChanged; selection
  popup ganha atalhos L/C/R/J para trechos selecionados
- Modo Roteiro: Courier New 12pt + alinhamento centralizado apenas no
  manuscrito (defaultTextOption + blockCountChanged singleShot)
- ProjectModel: projectType (book/screenplay), defaultManuscriptAlignment,
  defaultDrawerAlignment persistidos no JSON
- NewProjectFlow: radio Livro/Roteiro na tela de detalhes
- SettingsPanel 2 colunas; seção Menções (@) com toggle manuscrito
- MentionPopup: drill-down Root→Capítulos→Cenas, atalho @ch1-sc2,
  portal Capítulos ▶, ESC volta nível; inclusão de manuscrito opcional
- Theme: 30+ temas novos (Gruvbox família, Tokyo família, Meia-Noite
  família, Everforest, One Dark, Neve, Linho, Giz, Quartzo, Bambu,
  Mel, Areia, Palha, Açafrão, Manteiga, Carvão, Obsidiana, Nimbo,
  Profundo, Lavanda, Citrino, Ferrugem, Sálvia, Éter, Daguerreótipo,
  CRT Verde, Bauhaus, Nocturno, Papyrus, VHS, Wabi-Sabi, Manuscrito,
  Limão, Brutalismo)

+++ src/Exporter.h
@@ -22,7 +22,7 @@ public:
     // espaçamento e entrelinha.
     struct DocStyle {
         QString fontFamily;
-        int fontSize = 12;
+        double fontSize = 12;
         int lineHeightPercent = 100;
         bool firstLineIndent = true;
         int spacingBefore = 0;
